add back button

+++ Scoreboard/scripts/scoreboard.js
@@ -9,8 +9,8 @@ let teamTwo;
 let maxOvers=0;
 let timelinePrev= []
 let ballatNoball=0;
-let teamOneBat;
-let teamTwoBat;
+let teamOneBat=0;
+let teamTwoBat=0;
 let icons=['./images/Image0.jpg','./images/Image1.jpg','./images/Image2.jpg','./images/Image3.jpg','./images/Image4.jpg','./images/Image5.jpg'];
 let icon1;
 let icon2;
@@ -41,25 +41,25 @@ document.addEventListener('DOMContentLoaded',()=>{
     wicketButton.addEventListener("click",wicketFall);
 
     let zeroButton = document.getElementById('zeroButton');
-    zeroButton.addEventListener("click",zeroScored);
+    zeroButton.addEventListener("click",function(){updateRuns(0)});
 
     let oneButton = document.getElementById('oneButton');
-    oneButton.addEventListener("click",oneScored);
+    oneButton.addEventListener("click",function(){updateRuns(1)});
 
     let twoButton = document.getElementById('twoButton');
-    twoButton.addEventListener("click",twoScored);
+    twoButton.addEventListener("click",function(){updateRuns(2)});
 
     let threeButton = document.getElementById('threeButton');
-    threeButton.addEventListener("click",threeScored);
+    threeButton.addEventListener("click",function(){updateRuns(3)});
 
     let fourButton = document.getElementById('fourButton');
-    fourButton.addEventListener("click",fourScored);
+    fourButton.addEventListener("click",function(){updateRuns(4)});
     
     let fiveButton = document.getElementById('fiveButton');
-    fiveButton.addEventListener('click',fiveScored);
+    fiveButton.addEventListener('click',function(){updateRuns(5)});
 
     let sixButton = document.getElementById('sixButton');
-    sixButton.addEventListener("click",sixScored);
+    sixButton.addEventListener("click",function(){updateRuns(6)});
 
     let wideButton = document.getElementById('wideButton');
     wideButton.addEventListener("click",wideScored);
@@ -78,9 +78,7 @@ document.addEventListener('DOMContentLoaded',()=>{
 
     let undoButton = document.getElementById('undoButton');
     undoButton.addEventListener("click",undo);
-
-    // let welcomePage = document.getElementById('welcomePage');
-    // welcomePage.addEventListener('click',goToWelcome);
+    // alert("Tap on the start button to start scoring")
 }) 
 const showRuns = (e) =>{
     e.preventDefault();
@@ -90,6 +88,7 @@ const showRuns = (e) =>{
     numberWicket=0;
     numberRuns=0;
     timeline =[]
+    document.getElementById('thisOver').innerHTML = "This Over: "
     document.getElementById('h12').innerHTML="Score";
     document.getElementById('rr').innerHTML = "RunRate = "+ 0;
     document.getElementById('runs').innerHTML = numberRuns;
@@ -101,24 +100,25 @@ const showRuns = (e) =>{
     document.getElementById('textOvers').innerHTML = "Overs = ";
     document.getElementById('endInningsButton').disabled = false
     enable();
+    document.getElementById('startButton').disabled = true;
 }
 
     //ENABLING THE DISABLED BUTTON
-    const enable =() =>{
-        document.getElementById('startButton').disabled = false
+const enable =() =>{
+    console.log("Enable called")
         document.getElementById('wicketButton').disabled = false
         document.getElementById('zeroButton').disabled = false
         document.getElementById('oneButton').disabled = false
         document.getElementById('twoButton').disabled = false
         document.getElementById('threeButton').disabled = false
         document.getElementById('fourButton').disabled = false
+        document.getElementById('fiveButton').disabled = false
         document.getElementById('sixButton').disabled = false
         document.getElementById('wideButton').disabled = false
         document.getElementById('byeButton').disabled = false
         document.getElementById('legbyeButton').disabled = false
         document.getElementById('noballButton').disabled = false
         document.getElementById('undoButton').disabled = false
-        document.getElementById('fiveButton').disabled = false
     }
 //WHEN A WICKET FALLS INCREMENT BY 1
 const wicketFall = () =>{
@@ -126,95 +126,29 @@ const wicketFall = () =>{
     numberWicket++;
     numberBalls++;
     if (numberWicket<=10 && numberWicket>=0){
+    console.log("wicket "+ numberWicket)
     document.getElementById('wicket').innerHTML = numberWicket;
     timeline.push("W");
     completeOvers();
-    }
-    else{
-        document.getElementById('wicketButton').disabled = true
-        document.getElementById('zeroButton').disabled = true
-        document.getElementById('oneButton').disabled = true
-        document.getElementById('twoButton').disabled = true
-        document.getElementById('threeButton').disabled = true
-        document.getElementById('fourButton').disabled = true
-        document.getElementById('sixButton').disabled = true
-        document.getElementById('wideButton').disabled = true
-        document.getElementById('byeButton').disabled = true
-        document.getElementById('legbyeButton').disabled = true
-        document.getElementById('noballButton').disabled = true
-    }
+    // checkTarget();
+}    
 }
-//WHEN 0 RUNS SCORED INCRMENET BY 0
-const zeroScored =()=>{
+//0,1,2,3,4,5,6
+const updateRuns = (a) =>{
+    console.log(a)
     enable();
-    numberRuns=numberRuns+0;
+    numberRuns = numberRuns + a;
     numberBalls++;
-    timeline.push(0)
-    document.getElementById('runs').innerHTML = numberRuns;
-    completeOvers()
-}
-
-//WHEN 1 RUNS IS SCORED INCREMENT BY 1
-const oneScored = () => {
-    enable();
-    numberRuns = numberRuns + 1
-    numberBalls++;
-    timeline.push(1)
-    document.getElementById('runs').innerHTML = numberRuns;
+    timeline.push(a);
     completeOvers();
-}
-//WHEN 2 RUNS IS SCORED INCREMENT BY 2
-const twoScored = () => {
-    enable();
-    numberRuns = numberRuns + 2
-    numberBalls++
-    timeline.push(2)
     document.getElementById('runs').innerHTML = numberRuns;
-    completeOvers()
 
-}
-
-//WHEN 3 RUNS IS SCORED INCREMENT BY 3
-const threeScored = () => {
-    enable();
-    numberRuns = numberRuns + 3
-    numberBalls++;
-    timeline.push(3)
-    document.getElementById('runs').innerHTML = numberRuns;
-    completeOvers()
-}
-
-//WHEN 4 RUNS IS SCORED INCREMENT BY 4
-const fourScored = () => {
-    enable();
-    numberRuns = numberRuns + 4
-    numberBalls++;
-    timeline.push(4)
-    document.getElementById('runs').innerHTML = numberRuns;
-    completeOvers()
-}
-//When 5 Runs Scored
-const fiveScored = () =>{
-    enable();
-    numberRuns = numberRuns + 5
-    numberBalls++
-    timeline.push(5)
-    document.getElementById('runs').innerHTML = numberRuns;
-    completeOvers()
-}
-//WHEN 6 RUNS IS SCORED INCREMENT BY 6
-const sixScored = () => {
-    enable();
-    numberRuns = numberRuns + 6
-    numberBalls++;
-    timeline.push(6);
-    document.getElementById('runs').innerHTML = numberRuns;
-    completeOvers()
 }
 
 //WHEN A WIDE BALL IS BOWLED INCREASE RUN BY 1 BUT NOT BALL
 const wideScored = () => {
     enable();
+    document.getElementById('wideButton').disabled = true
     document.getElementById('legbyeButton').disabled = true
     document.getElementById('noballButton').disabled = true
     document.getElementById('byeButton').disabled = true
@@ -228,6 +162,7 @@ const wideScored = () => {
 //WHEN A BYE IS SCORED
 const byeScored = () =>{
     enable();
+    document.getElementById('byeButton').disabled = true
     document.getElementById('legbyeButton').disabled = true
     document.getElementById('noballButton').disabled = true
     document.getElementById('wideButton').disabled = true
@@ -239,19 +174,21 @@ const byeScored = () =>{
 //WHEN LEGBYE SCORED
 const legbyeScored = () =>{
     enable();
+    document.getElementById('legbyeButton').disabled = true
     document.getElementById('byeButton').disabled = true
     document.getElementById('noballButton').disabled = true
     document.getElementById('wideButton').disabled = true
     document.getElementById('wicketButton').disabled = true
     document.getElementById('startButton').disabled = true
     document.getElementById('endInningsButton').disabled = true
-    timeline.push("lb");
+    timeline.push("Lb");
 
 }
 //WHEN A NO BALL IS BOWLED
 const noBall = () =>{
     enable();
     ballatNoball = numberBalls;
+    document.getElementById('noballButton').disabled = true
     document.getElementById('byeButton').disabled = true
     document.getElementById('legbyeButton').disabled = true
     document.getElementById('wideButton').disabled = true
@@ -317,7 +254,6 @@ const undo = () =>{
         numberWicket = numberWicket -1;
         document.getElementById('wicket').innerHTML = numberWicket;
     }
-    //Current
     else if(undoValue2 === "W" && undoValue === 'Nb'){
         timeline.pop();
         timeline.pop();
@@ -326,7 +262,12 @@ const undo = () =>{
         document.getElementById('wicket').innerHTML = numberWicket;
         document.getElementById('wicketButton').value = "Wicket";
     }
-    //WORKING
+    else if(typeof(undoValue2)==="number" && (undoValue == "B"||undoValue === "Lb")){
+        numberRuns = numberRuns - undoValue2;
+        numberBalls = numberBalls -1;
+        timeline.pop();
+        timeline.pop();
+    }
     else if(typeof(undoValue2)==="number" && (undoValue != "Wd"||undoValue != "Nb")){
         numberRuns = numberRuns - undoValue2;
         numberBalls = numberBalls -1;
@@ -352,7 +293,11 @@ else if(numberBalls===0 && timeline[1] ==='W' && (timeline[0] ==='Wd' || timelin
     document.getElementById('wicketButton').value = "Wicket"
 }
 else{
- 
+    if(numberBalls ===0 && numberOvers ===0){
+        alert('Nothing to undo')
+}
+
+else{
     let undo0BallRuns = timelinePrev[timelinePrev.length -1 ];
     console.log(typeof(undo0BallRuns) + undo0BallRuns);
     numberRuns = numberRuns -undo0BallRuns;
@@ -362,8 +307,8 @@ else{
     numberOvers--;
     document.getElementById('overs').innerHTML = numberOvers;
 }
+}
     timelineUpdate();
-    document.getElementById('runs').innerHTML = numberRuns;
     completeOvers();
 }
 //Timeline Updation
@@ -373,12 +318,13 @@ const timelineUpdate=()=>{
 }
 //Complete Over Function with overs and balls
  const completeOvers = () =>{
+    document.getElementById('runs').innerHTML = numberRuns;
     if(ballatNoball === numberBalls - 1){
         document.getElementById('wicketButton').value = 'Wicket';
     }
-    if (numberBalls<=5){
+    if (numberBalls<=5 && numberBalls>=0){
         document.getElementById('balls').innerHTML = numberBalls;
-        }
+    }
     else {
         numberOvers++;
         document.getElementById('overs').innerHTML = numberOvers;
@@ -388,7 +334,6 @@ const timelineUpdate=()=>{
     timelineUpdate();
 
     if(numberBalls === 0 && (timeline[0] != 'Wd' && timeline[0] !='Nb' )){
-        console.log("Hweloo")
         timelinePrev = [...timeline]
         timeline = []
     }
@@ -398,7 +343,7 @@ const timelineUpdate=()=>{
     else{
         document.getElementById('rr').innerHTML = "RunRate = "+(numberRuns/(numberOvers*6+numberBalls)*6).toFixed(1);
 }
-    if(numberOvers === maxOvers && target === 0){
+    if((numberOvers === maxOvers && target === 0 )||(numberWicket ===10 && target===0)){
         endInnings();
     }
     else{
@@ -418,19 +363,25 @@ const checkTarget = ()=>{
         }
         if(numberOvers === maxOvers ){
             if(numberRuns<target-1){
+                console.log("Hi")
                 alert(teamOne + " is Winner")
                 goToWelcome();
-            localStorage.clear();
+                localStorage.clear();
 
             }
                 if(numberRuns === target -1 ){
                     alert("Match Drawn")
                     goToWelcome();
-            localStorage.clear();
-
+                    localStorage.clear();
                 }
             }
-        if(numberWicket ===10){
+            if(numberRuns === target -1 && numberWicket===10){
+                alert("Match Drawn")
+                goToWelcome();
+                localStorage.clear();
+            }
+        if(numberWicket ===10 && numberOvers !=maxOvers && numberRuns<target-1){
+            console.log("Hello")
             alert(teamOne + " is Winner")
             goToWelcome();
             localStorage.clear();
@@ -453,14 +404,19 @@ const checkTarget = ()=>{
                 if(numberRuns === target -1 ){
                     alert("Match Drawn")
                     goToWelcome();
-                     localStorage.clear();
+                    localStorage.clear();
 
                 }
             }
-        if(numberWicket ===10){
+            if(numberRuns === target -1 && numberWicket===10){
+                alert("Match Drawn")
+                goToWelcome();
+                localStorage.clear();
+            }
+        if(numberWicket ===10 && numberOvers!=maxOvers && numberRuns<target-1){
             alert(teamTwo + " is Winner")
-            goToWelcome();
-            localStorage.clear();
+             goToWelcome();
+             localStorage.clear();
 
         }
     }
@@ -468,6 +424,6 @@ const checkTarget = ()=>{
 }
 //Redirection To Welcomwe Page
 const goToWelcome =() =>{
-    alert("You Want to go to Match Setup? All Information would be lost");
+    alert("You are being redirected to welcome page.All Information would be lost");
     window.location.replace("welcome.html")
 }
